Add method for loading the config
This takes a file and reads it into a struct. In case of an error it
always returns enduser friendly error messages. All errors and two
common success cases are tested.

Parsing the file as a value and converting it to a table is unwrapped
because every file that parses to a toml value is a valid table as far
as I know. If this is wrong it should be changed.