Fix comment being overwritten
Due to the way it worked until now it was not possible to keep comments
on styles that already had settings. Now it has been changed by always
updating the settings for all styles. This will always add the currently
available options to the comments. However values are not reset,
but if a key doesn't exist anymore it will be removed.

The biggest drawback is that the user will not be notified when an
option does not exist anymore, because the custom changes the user made
might not be in the `settings_options`.
It would be possible to change removed values back to the default for
settings like dropdowns, because the user can only pick from fixed
values here.